test: drop the façade tests that restate another one
Collapse the two files that were organised by circumstance rather than by
subject: test_node_edge_cases.py's tests go to the endpoint each is about, and
test_node_create_validation.py joins test_node_create.py, so POST's behaviour
and POST's refusals are read together.

Remove the tests another one already makes. Two were subsumed outright: a write
response's mcp_tools claim, and the graph-versus-row comparison -- which cannot
fail at all, since flow_data rebuilds its nodes from the rows (ADR-0049). The
rest were a second assertion on an identical request, or a narrower case of the
test beside it; where one of those carried a claim of its own it moves to the
test that stays.

Stop sweeping the auth gates across all three verbs. One view class serves the
three node endpoints and every gate on it turns on the credential: the
read-only key gate gates on safe-versus-unsafe methods, the OAuth resource
scope maps all three to chatbots:write, and ChatbotCompositionPermission never
reads the verb. Only the model-permission test is exercised per verb, which is
where the verb could have changed the answer -- the stock DjangoModelPermissions
map would have demanded delete_experiment for DELETE.
