+++ apps/api/v2/pipeline_edit/tests/test_node_create.py
@@ -1,8 +1,20 @@
-"""POST /api/v2/chatbots/{id}/pipeline/nodes/ (#4140)."""
+"""POST /api/v2/chatbots/{id}/pipeline/nodes/ (#4140, spec §6.2).
+
+What the endpoint produces, and what it refuses. The rule the refusals encode: a structurally-sound
+node always persists, even when it is semantically incomplete, so an agent can build a graph a piece
+at a time. What does *not* persist is a request naming something that does not exist — a node type,
+or a resource id.
+"""
+
+from unittest.mock import Mock, patch
 
 import pytest
 
 from apps.pipelines.models import Node
+from apps.utils.factories.documents import CollectionFactory
+from apps.utils.factories.experiment import SourceMaterialFactory
+from apps.utils.factories.service_provider_factories import LlmProviderFactory
+from apps.utils.factories.team import TeamWithUsersFactory
 
 from .conftest import nodes_url
 
@@ -27,52 +39,72 @@ def test_create_adds_a_node_with_a_server_assigned_id(client, chatbot, llm):
 
 
 @pytest.mark.django_db()
-def test_create_persists_the_node_types_defaults(client, chatbot):
-    """The node class is the only place the defaults live, and `update_nodes_from_data` stores
-    params verbatim -- so unless they are materialized here, a node created through the API reads
-    back from /inspect/ as the handful of keys the client happened to send."""
+def test_create_fills_in_the_node_types_defaults(client, chatbot):
+    """`type` alone has to be enough to add a node.
+
+    The node class is the only place the defaults live, and `update_nodes_from_data` stores params
+    verbatim -- so unless they are materialized here, a node created through the API reads back from
+    /inspect/ as the handful of keys the client happened to send. `name` is required on every type
+    and has no default, so the server supplies the node id, which is what the builder writes when a
+    node is dragged onto the canvas; the label is the type's display name, since the builder shows
+    it and blank would read as nothing at all.
+    """
     response = client.post(nodes_url(chatbot), {"type": "LLMResponseWithPrompt"}, format="json")
 
     assert response.status_code == 201, response.content
-    node = Node.objects.get(pipeline=chatbot.pipeline, flow_id=response.json()["node"]["node_id"])
+    node_id = response.json()["node"]["node_id"]
+    node = Node.objects.get(pipeline=chatbot.pipeline, flow_id=node_id)
     assert node.params["history_type"] == "global"
     assert node.params["max_history_length"] == 10
-
-
-@pytest.mark.django_db()
-def test_create_names_the_node_after_its_id(client, chatbot):
-    """`name` is required on every node type and has no default, so the server has to supply one.
-    It uses the node id, which is what the builder writes when a node is dragged onto the canvas."""
-    response = client.post(nodes_url(chatbot), {"type": "LLMResponseWithPrompt"}, format="json")
-
-    node_id = response.json()["node"]["node_id"]
-    assert Node.objects.get(pipeline=chatbot.pipeline, flow_id=node_id).params["name"] == node_id
-
-
-@pytest.mark.django_db()
-def test_create_accepts_a_name_over_the_default(client, chatbot):
-    response = client.post(
-        nodes_url(chatbot), {"type": "LLMResponseWithPrompt", "params": {"name": "classifier"}}, format="json"
-    )
-
-    node_id = response.json()["node"]["node_id"]
-    assert Node.objects.get(pipeline=chatbot.pipeline, flow_id=node_id).params["name"] == "classifier"
-
-
-@pytest.mark.django_db()
-def test_create_labels_the_node_with_its_types_display_name(client, chatbot):
-    """The builder shows a node's label, so a node created without one has to arrive with
-    something readable rather than blank."""
-    response = client.post(nodes_url(chatbot), {"type": "LLMResponseWithPrompt"}, format="json")
-
+    assert node.params["name"] == node_id
     assert response.json()["node"]["label"] == "LLM"
 
 
 @pytest.mark.django_db()
-def test_create_accepts_a_label(client, chatbot):
-    response = client.post(nodes_url(chatbot), {"type": "LLMResponseWithPrompt", "label": "Classify"}, format="json")
+def test_create_takes_a_label_and_a_name_over_the_defaults(client, chatbot):
+    response = client.post(
+        nodes_url(chatbot),
+        {"type": "LLMResponseWithPrompt", "label": "Classify", "params": {"name": "classifier"}},
+        format="json",
+    )
 
-    assert response.json()["node"]["label"] == "Classify"
+    assert response.status_code == 201, response.content
+    body = response.json()
+    assert body["node"]["label"] == "Classify"
+    assert Node.objects.get(pipeline=chatbot.pipeline, flow_id=body["node"]["node_id"]).params["name"] == "classifier"
+
+
+@pytest.mark.django_db()
+def test_a_colliding_node_id_is_redrawn(client, chatbot):
+    """``apply_pipeline_patch`` treats an add whose id already exists as a no-op, so a clash would
+    answer 201 while describing the node that was already there."""
+    created = client.post(nodes_url(chatbot), {"type": "CodeNode"}, format="json")
+    taken = created.json()["node"]["node_id"]
+    collision = taken.removeprefix("CodeNode-")
+
+    # The first draw repeats the id already in the graph, the second is free.
+    draws = [Mock(hex=f"{collision}0000000"), Mock(hex="abcde" + "0" * 27)]
+    with patch("apps.api.v2.pipeline_edit.graph_editor.uuid4", side_effect=draws):
+        response = client.post(nodes_url(chatbot), {"type": "CodeNode"}, format="json")
+
+    assert response.status_code == 201, response.content
+    assert response.json()["node"]["node_id"] == "CodeNode-abcde"
+    assert Node.objects.filter(pipeline=chatbot.pipeline, type="CodeNode").count() == 2
+
+
+@pytest.mark.django_db()
+def test_an_id_source_that_only_collides_still_answers(client, chatbot):
+    """The redraw is bounded: an id source stuck on one value must fall back to a longer id rather
+    than spin inside the pipeline row lock until the request times out."""
+    created = client.post(nodes_url(chatbot), {"type": "CodeNode"}, format="json")
+    collision = created.json()["node"]["node_id"].removeprefix("CodeNode-")
+
+    with patch("apps.api.v2.pipeline_edit.graph_editor.uuid4", return_value=Mock(hex=f"{collision}{'0' * 27}")):
+        response = client.post(nodes_url(chatbot), {"type": "CodeNode"}, format="json")
+
+    assert response.status_code == 201, response.content
+    assert response.json()["node"]["node_id"] == f"CodeNode-{collision}{'0' * 27}"
+    assert Node.objects.filter(pipeline=chatbot.pipeline, type="CodeNode").count() == 2
 
 
 @pytest.mark.django_db()
@@ -123,6 +155,160 @@ def test_create_moves_the_output_node_without_rewriting_what_it_holds(client, ch
     end = _end_node(chatbot.pipeline)
     assert end.params == {"name": "end"}
     assert end.label == ""
+
+
+@pytest.mark.django_db()
+def test_an_unknown_node_type_is_refused(client, chatbot):
+    """404, the same answer /pipeline/nodes/{type}/ gives: the name is of a type that does not
+    exist, whichever way the client happened to name it. The valid ones come back with it."""
+    response = client.post(nodes_url(chatbot), {"type": "Frobnicator"}, format="json")
+
+    assert response.status_code == 404, response.content
+    assert "LLMResponseWithPrompt" in response.json()["valid_types"]
+    assert not chatbot.pipeline.node_set.filter(type="Frobnicator").exists()
+
+
+@pytest.mark.django_db()
+def test_a_server_managed_node_type_is_refused(client, chatbot):
+    """Start and End are created with the pipeline and are not something a client may add — the
+    same refusal /pipeline/nodes/{type}/ already gives for them."""
+    response = client.post(nodes_url(chatbot), {"type": "StartNode"}, format="json")
+
+    assert response.status_code == 404, response.content
+    assert "managed by the server" in response.json()["detail"]
+
+
+@pytest.mark.django_db()
+def test_a_body_with_no_type_is_refused(client, chatbot):
+    response = client.post(nodes_url(chatbot), {"params": {"name": "orphan"}}, format="json")
+
+    assert response.status_code == 400, response.content
+    assert "type" in response.json()
+
+
+@pytest.mark.django_db()
+def test_a_client_supplied_node_id_is_refused(client, chatbot):
+    """Ids are the server's to assign (W5): honouring a client's would let two nodes collide."""
+    response = client.post(nodes_url(chatbot), {"type": "LLMResponseWithPrompt", "node_id": "mine-1"}, format="json")
+
+    assert response.status_code == 400, response.content
+    assert "server" in str(response.json()["node_id"]).lower()
+    assert not Node.objects.filter(pipeline=chatbot.pipeline, flow_id="mine-1").exists()
+
+
+@pytest.mark.django_db()
+def test_an_unrecognised_body_key_is_refused(client, chatbot):
+    response = client.post(nodes_url(chatbot), {"type": "LLMResponseWithPrompt", "colour": "red"}, format="json")
+
+    assert response.status_code == 400, response.content
+    assert "colour" in response.json()
+
+
+@pytest.mark.django_db()
+def test_an_unrecognised_param_is_dropped(client, chatbot):
+    response = client.post(
+        nodes_url(chatbot), {"type": "LLMResponseWithPrompt", "params": {"tempreture": 0.5}}, format="json"
+    )
+
+    assert response.status_code == 201, response.content
+    node_id = response.json()["node"]["node_id"]
+    assert "tempreture" not in response.json()["node"]["params"]
+    assert "tempreture" not in Node.objects.get(pipeline=chatbot.pipeline, flow_id=node_id).params
+
+
+@pytest.mark.django_db()
+def test_a_missing_required_param_persists_and_is_reported(client, chatbot):
+    """Lenient on structure: the node lands so the next call can fill it in, and the gap shows up
+    in the errors report the publish gate rejects on."""
+    response = client.post(nodes_url(chatbot), {"type": "LLMResponseWithPrompt"}, format="json")
+
+    assert response.status_code == 201, response.content
+    body = response.json()
+    node_id = body["node"]["node_id"]
+    assert Node.objects.filter(pipeline=chatbot.pipeline, flow_id=node_id).exists()
+    assert body["pipeline_valid"] is False
+    assert "llm_provider_id" in body["pipeline_errors"]["node"][node_id]
+
+
+@pytest.mark.django_db()
+def test_a_reference_to_another_teams_resource_is_refused(client, chatbot):
+    """Indistinguishable from a nonexistent id on purpose: telling the two apart would answer
+    whether the id exists in some other team."""
+    elsewhere = LlmProviderFactory.create(team=TeamWithUsersFactory.create(), type="openai")
+
+    response = client.post(
+        nodes_url(chatbot),
+        {"type": "LLMResponseWithPrompt", "params": {"llm_provider_id": elsewhere.id}},
+        format="json",
+    )
+
+    assert response.status_code == 400, response.content
+    assert "llm_provider_id" in response.json()["params"]
+    assert not chatbot.pipeline.node_set.filter(type="LLMResponseWithPrompt").exists()
+
+
+@pytest.mark.django_db()
+def test_a_reference_to_the_teams_own_resource_is_accepted(client, chatbot, team):
+    """Guards the guard: the refusal above has to be about the team boundary, not about the
+    reference check refusing every id it is shown."""
+    material = SourceMaterialFactory.create(team=team)
+
+    response = client.post(
+        nodes_url(chatbot),
+        {"type": "LLMResponseWithPrompt", "params": {"source_material_id": material.id}},
+        format="json",
+    )
+
+    assert response.status_code == 201, response.content
+
+
+@pytest.mark.django_db()
+def test_a_duplicate_node_name_persists_and_is_reported(client, chatbot, llm):
+    """`name` is how one node reaches another's output, so a clash breaks the pipeline -- but it is
+    structural, so it is reported rather than refused."""
+    provider, model = llm
+    params = {"llm_provider_id": provider.id, "llm_provider_model_id": model.id, "name": "classifier"}
+
+    first = client.post(nodes_url(chatbot), {"type": "LLMResponseWithPrompt", "params": params}, format="json")
+    second = client.post(nodes_url(chatbot), {"type": "LLMResponseWithPrompt", "params": params}, format="json")
+
+    assert (first.status_code, second.status_code) == (201, 201), second.content
+    clashing = second.json()["pipeline_errors"]["node"]
+    assert [error["name"] for error in clashing.values()] == ["All node names must be unique"] * 2
+
+
+@pytest.mark.django_db()
+def test_a_list_valued_reference_is_checked_per_entry(client, chatbot, team):
+    """`collection_index_ids` holds a list, so the check has to look at every entry rather than at
+    the list as a whole."""
+    ours = CollectionFactory.create(team=team, is_index=True)
+    theirs = CollectionFactory.create(team=TeamWithUsersFactory.create(), is_index=True)
+
+    response = client.post(
+        nodes_url(chatbot),
+        {"type": "LLMResponseWithPrompt", "params": {"collection_index_ids": [ours.id, theirs.id]}},
+        format="json",
+    )
+
+    assert response.status_code == 400, response.content
+    assert str(theirs.id) in response.json()["params"]["collection_index_ids"]
+    assert str(ours.id) not in response.json()["params"]["collection_index_ids"]
+
+
+@pytest.mark.django_db()
+def test_a_malformed_custom_action_reference_is_refused(client, chatbot):
+    """`custom_actions` entries are the composite "{action_id}:{operation_id}" strings the server
+    hands out, and `Node.update_from_params` splits them on the colon -- so a value that is not one
+    would be a 500 rather than a rejected write if it got as far as being saved."""
+    response = client.post(
+        nodes_url(chatbot),
+        {"type": "LLMResponseWithPrompt", "params": {"custom_actions": ["not-a-reference"]}},
+        format="json",
+    )
+
+    assert response.status_code == 400, response.content
+    assert "custom_actions" in response.json()["params"]
+    assert not chatbot.pipeline.node_set.filter(type="LLMResponseWithPrompt").exists()
 
 
 def _place(pipeline, start: int, end: int) -> None:
