+++ apps/api/v2/pipeline_edit/tests/test_node_param_types.py
@@ -1,13 +1,14 @@
 """What the node model settles about a param's value, and what it only reports (#4140).
 
 A loosely typed value is normalised on the way in; one that cannot be parsed is stored and reported
-in ``pipeline_errors``. Only a reference the team cannot reach is refused; an undeclared param is
-dropped.
+in ``pipeline_errors``. Only a reference the team cannot reach — or one of a shape the check cannot
+read — is refused.
 """
 
 import pytest
 
 from apps.pipelines.models import Node
+from apps.utils.factories.documents import CollectionFactory
 
 from .conftest import node_url, nodes_url
 
@@ -22,6 +23,7 @@ from .conftest import node_url, nodes_url
         pytest.param("CodeNode", {"code": ["print(1)"]}, "root", id="array-for-string"),
         pytest.param("RouterNode", {"keywords": "one"}, "keywords", id="string-for-array"),
         pytest.param("RouterNode", {"keywords": [1, 2]}, "keywords", id="ints-for-string-array"),
+        pytest.param("LLMResponseWithPrompt", {"max_history_length": "loads"}, "max_history_length", id="word-for-int"),
     ],
 )
 def test_a_param_of_an_unparseable_type_persists_and_is_reported(client, chatbot, node_type, params, error_key):
@@ -35,9 +37,23 @@ def test_a_param_of_an_unparseable_type_persists_and_is_reported(client, chatbot
 
 
 @pytest.mark.django_db()
-def test_a_scalar_reference_sent_as_a_list_is_refused(client, chatbot, llm):
-    """Nothing type-checks params before the reference check now, so it has to handle an unhashable
-    value itself rather than raising on it."""
+def test_patch_stores_an_unparseable_param_and_reports_it(client, chatbot):
+    created = client.post(nodes_url(chatbot), {"type": "CodeNode"}, format="json")
+    node_id = created.json()["node"]["node_id"]
+
+    response = client.patch(node_url(chatbot, node_id), {"params": {"code": 123}}, format="json")
+
+    assert response.status_code == 200, response.content
+    body = response.json()
+    assert body["pipeline_valid"] is False
+    assert body["pipeline_errors"]["node"][node_id]["root"]
+    assert Node.objects.get(pipeline=chatbot.pipeline, flow_id=node_id).params["code"] == 123
+
+
+@pytest.mark.django_db()
+def test_a_list_sent_for_a_scalar_reference_is_refused(client, chatbot, llm):
+    """Nothing type-checks params before the reference check, so it has to handle a value of the
+    wrong shape itself rather than raising on it. Here that means an unhashable one."""
     provider, model = llm
 
     response = client.post(
@@ -55,28 +71,19 @@ def test_a_scalar_reference_sent_as_a_list_is_refused(client, chatbot, llm):
 
 
 @pytest.mark.django_db()
-def test_an_unparseable_param_leaves_the_pipeline_readable(client, chatbot):
-    client.post(nodes_url(chatbot), {"type": "CodeNode", "params": {"code": 123}}, format="json")
+def test_a_scalar_sent_for_a_list_valued_reference_is_refused(client, chatbot, team):
+    """The other shape: a bare id where a list belongs. Read as one unreachable value rather than
+    iterated over as if it were a list."""
+    ours = CollectionFactory.create(team=team, is_index=True)
 
-    inspect = client.get(f"/api/v2/chatbots/{chatbot.public_id}/inspect/")
-    assert inspect.status_code == 200, inspect.content
+    response = client.post(
+        nodes_url(chatbot),
+        {"type": "LLMResponseWithPrompt", "params": {"collection_index_ids": ours.id + 1000}},
+        format="json",
+    )
 
-    follow_up = client.post(nodes_url(chatbot), {"type": "CodeNode"}, format="json")
-    assert follow_up.status_code == 201, follow_up.content
-
-
-@pytest.mark.django_db()
-def test_patch_stores_an_unparseable_param_and_reports_it(client, chatbot):
-    created = client.post(nodes_url(chatbot), {"type": "CodeNode"}, format="json")
-    node_id = created.json()["node"]["node_id"]
-
-    response = client.patch(node_url(chatbot, node_id), {"params": {"code": 123}}, format="json")
-
-    assert response.status_code == 200, response.content
-    body = response.json()
-    assert body["pipeline_valid"] is False
-    assert body["pipeline_errors"]["node"][node_id]["root"]
-    assert Node.objects.get(pipeline=chatbot.pipeline, flow_id=node_id).params["code"] == 123
+    assert response.status_code == 400, response.content
+    assert "collection_index_ids" in response.json()["params"]
 
 
 @pytest.mark.django_db()
